se realiza punto f). Creamos/modificamos Controller, Repository y Service de Carrera, modificamos atributos de entity a Long

+++ src/main/java/org/example/grupo20integrador3/entities/Carrera.java
@@ -31,7 +31,7 @@ public class Carrera implements Serializable {
         this.inscriptos = new ArrayList<EstudianteCarrera>();
     }
 
-    public int getInscriptos() {
-        return inscriptos.size();
+    public Long getInscriptos() {
+        return (long) inscriptos.size();
     }
 }
